requirements.txt
streamlit 
langchain 
openai 
chromadb 
python-dotenv 
PyPDF2 
tiktoken 
